+++ amundsen_application/static/js/pages/TableDetailPage/ColumnList/ColumnStats/index.spec.tsx
