+++ amundsen_application/static/js/pages/TableDetailPage/ColumnList/ColumnStats/styles.scss
