+++ amundsen_application/static/js/pages/TableDetailPage/ColumnListItem/index.tsx
@@ -11,7 +11,7 @@ import { TableColumn, RequestMetadataType } from 'interfaces';
 import { OpenRequestAction } from 'ducks/notification/types';
 
 import EditableSection from 'components/common/EditableSection';
-import ColumnStats from '../ColumnStats';
+import ColumnStats from '../ColumnList/ColumnStats';
 import ColumnDescEditableText from '../ColumnDescEditableText';
 import ColumnType from '../ColumnList/ColumnType';
